+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+- Added support for video embeds: currently only supports `{{#ev:youtube}}` and `{{#ev:vimeo}}`.
 - Changed images to be displayed inline by default when no float is given.
 - Fixed unused parameters being substituted with random numbers instead of their default values.
 - Fixed images being erroneously parsed as regular links.